requirements.txt
fpdf
Pillow